task06 - исправил названия папок
