--- libraryProject/src/main/java/org/example/LibraryAppApplication.java
@@ -1,11 +1,0 @@
-package org.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LibraryAppApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(LibraryAppApplication.class, args);
-    }
-}
