--- libraryProject/src/main/java/org/example/dao/AuthorDao.java
@@ -1,21 +1,0 @@
-package org.example.dao;
-
-import org.example.model.Author;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface AuthorDao {
-    List<Author> findAll();
-
-    Optional<Author> findById(Long id);
-
-    void insert(Author author);
-
-    void update(Author author);
-
-    /**
-     * Удаляет автора по id. Возвращает true, если был удалён, иначе false.
-     */
-    boolean deleteById(Long id);
-}
