--- libraryProject/src/main/java/org/example/model/Author.java
@@ -1,11 +1,0 @@
-package org.example.model;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Author {
-    private Long id;
-    private String name;
-}
