--- libraryProject/src/main/java/org/example/model/Book.java
@@ -1,14 +1,0 @@
-package org.example.model;
-
-import lombok.*;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Book {
-    private Long id;
-    private String title;
-    private Genre genre;
-    private List<Author> authors;
-}
