--- libraryProject/src/main/java/org/example/model/Genre.java
@@ -1,11 +1,0 @@
-package org.example.model;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Genre {
-    private Long id;
-    private String name;
-}
