--- libraryProject/src/main/java/org/example/service/BookService.java
@@ -1,17 +1,0 @@
-package org.example.service;
-
-import org.example.model.Book;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface BookService {
-    List<Book> findAll();
-    Optional<Book> findById(Long id);
-    void insert(Book book);
-    void update(Book book);
-    /**
-     * Удаляет книгу по id. Возвращает true, если была удалена, иначе false.
-     */
-    boolean deleteById(Long id);
-}
