+++ libraryProject/src/main/java/ru/otus/vseroev/LibraryAppApplication.java
@@ -1,0 +1,11 @@
+package ru.otus.vseroev;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LibraryAppApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(LibraryAppApplication.class, args);
+    }
+}
