+++ libraryProject/src/main/java/ru/otus/vseroev/dao/AuthorDao.java
@@ -1,0 +1,21 @@
+package ru.otus.vseroev.dao;
+
+import ru.otus.vseroev.model.Author;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface AuthorDao {
+    List<Author> findAll();
+
+    Optional<Author> findById(Long id);
+
+    void insert(Author author);
+
+    void update(Author author);
+
+    /**
+     * Удаляет автора по id. Возвращает true, если был удалён, иначе false.
+     */
+    boolean deleteById(Long id);
+}
