+++ libraryProject/src/main/java/ru/otus/vseroev/model/Author.java
@@ -1,0 +1,11 @@
+package ru.otus.vseroev.model;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Author {
+    private Long id;
+    private String name;
+}
