+++ libraryProject/src/main/java/ru/otus/vseroev/model/Book.java
@@ -1,0 +1,14 @@
+package ru.otus.vseroev.model;
+
+import lombok.*;
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Book {
+    private Long id;
+    private String title;
+    private Genre genre;
+    private List<Author> authors;
+}
