+++ libraryProject/src/main/java/ru/otus/vseroev/model/Genre.java
@@ -1,0 +1,11 @@
+package ru.otus.vseroev.model;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Genre {
+    private Long id;
+    private String name;
+}
